CLI update v2
